@tomaarsen 2024-11-19
With the `dict.fromkeys` (Thanks @Wauplin for the suggestion), the performance hit is reduced by a bit.

### This PR (new)
Time for sampling 1 batch:
```
Time: 0.23999476432800293
Time: 0.2646820545196533
Time: 0.2566261291503906
Time: 0.2643446922302246
Time: 0.2761116027832031
```
(down from ~0.31s avg)

Time for sampling 1k batches:
```
Time: 0.604529619216919
Time: 0.6064455509185791
Time: 0.6098651885986328
Time: 0.6006977558135986
Time: 0.5981855392456055
```
(down from ~0.635s avg)

The 1 batch case actually seems faster than the baseline - awesome!

- Tom Aarsen