@antigregory 2024-11-19
Thanks for the PR @tomaarsen!
As you said, the performance & memory overheads seem very reasonable.

However, I'm not sure to have understood what alternative you were benchmarking in the paragraph **Deleting while iterating** (but it's not very important given the poor results of this alternative).