@tomaarsen 2024-11-19
> However, I'm not sure to have understood what alternative you were benchmarking in the paragraph Deleting while iterating (but it's not very important given the poor results of this alternative).

I didn't explain it very well indeed, apologies. I'll clarify (regardless of the horrible results, hah!)
In this PR, I'm performing
```python
            for index in batch_indices:
                del remaining_indices[index]
```
after that batch has been sampled.

I can also do this `del remaining_indices[index]` during the iteration itself. E.g.:
```python
                batch_indices.append(index)
                del remaining_indices[index]
                if len(batch_indices) == self.batch_size:
                    yield batch_indices
                    break
```
The idea being that you then don't need to iterate over the `batch_indices` *again*.

But dictionaries don't like it if you update them while you're iterating over them. So, you have to iterate over a copy or you'll get an error. So, instead of 
```python
            for index in remaining_indices:
```
I tried
```python
            for index in list(remaining_indices.keys()):
```
Needless to say - that didn't help.

- Tom Aarsen