@tomaarsen 2024-11-20
I'll merge this, but it won't be fully fixed until https://github.com/huggingface/accelerate/pull/3246 is also merged and released. 

- Tom Aarsen